--- central/client/client.html
@@ -1,71 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-    <head>
-        <title>Some Random Chat</title>
-        <script src="jquery-1.9.1.min.js"></script>
-        <script src="bootstrap/tether-1.3.3/dist/js/tether.min.js"></script>
-        <script src="bootstrap/js/bootstrap.min.js"></script>
-        <script src="script.js"></script>
-        <link rel="stylesheet" type="text/css" href="bootstrap/css/bootstrap.min.css">
-    </head>
-    <body>
-
-        <div class="container">
-            <div class="row" style="text-align: center;">
-                <div class="col-md-12">
-                    <h1>Some Random Chat</h1>
-                </div>
-            </div>
-
-            <br>
-
-            <form id="nameForm" name="nameForm" novalidate>
-                <div class="form-group">
-                    <input type="email" class="form-control" id="name" aria-describedby="nameHelp" placeholder="Your name?">
-                    <small id="nameHelp" class="form-text text-muted">Your name lets people to know you are here!</small>
-                </div>
-
-                <button type="submit" class="btn btn-primary" style="margin-left: 50px;">Enter</button>
-            </form>
-
-            <div class="row hidden-xs-up" id="chatSection">
-                <div class="col-md-6">
-                    <form id="chatsForm" name="chatsForm" novalidate>
-                        <div class="form-group">
-                            <label for="chats">Select a chat</label>
-                            <select class="form-control" id="chats"></select>
-                        </div>
-
-                        <button type="submit" class="btn btn-primary" style="margin-left: 50px;">Go!</button>
-                    </form>
-                </div>
-                <div class="col-md-6">
-                    <form id="chatNameForm" name="chatNameForm" novalidate>
-                        <div class="form-group">
-                            <label for="chatName">Or create one</label>
-                            <input type="text" class="form-control" id="chatName"/>
-                        </div>
-
-                        <button type="submit" class="btn btn-primary" style="margin-left: 50px;">Go!</button>
-                    </form>
-                </div>
-            </div>
-
-            <form id="messageForm" class="hidden-xs-up" name="messageForm" novalidate>
-                <div class="form-group">
-                    <label for="message">Message</label>
-                    <textarea class="form-control" id="message" rows="3" cols="42"></textarea>
-                </div>
-
-                <button type="submit" class="btn btn-primary" style="margin-left: 50px;">Send</button>
-            </form>
-
-            <br>
-            <br>
-
-            <div class="row" id="chat"></div>
-        </div>
-
-    </body>
-
-</html>
