--- discovery/client/client.html
@@ -1,35 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-    <head>
-        <title>Some Random Chat</title>
-        <script src="http://cdn.peerjs.com/0.3/peer.js"></script>
-        <script src="jquery-1.9.1.min.js"></script>
-        <script src="bootstrap/tether-1.3.3/dist/js/tether.min.js"></script>
-        <script src="bootstrap/js/bootstrap.min.js"></script>
-        <script src="script.js"></script>
-        <link rel="stylesheet" type="text/css" href="bootstrap/css/bootstrap.min.css">
-    </head>
-
-    <body>
-        <div class="row">
-            <form id="peerForm" name="peerForm" novalidate>
-                <div class="form-group">
-                    <label for="peers">Select a peer</label>
-                    <select class="form-control" id="peers"></select>
-                </div>
-
-                <button type="submit" class="btn btn-primary" style="margin-left: 50px;">Go!</button>
-            </form>
-        </div>
-
-        <form id="messageForm" class="hidden-xs-up" name="messageForm" novalidate>
-            <div class="form-group">
-                <label for="message">Message</label>
-                <textarea class="form-control" id="message" rows="3" cols="42"></textarea>
-            </div>
-
-            <button type="submit" class="btn btn-primary" style="margin-left: 50px;">Send</button>
-        </form>
-
-    </body>
-</html>
